requirements.txt
discord==1.0.1
discord.py==1.7.1
discord-py-slash-command==1.1.2
requests==2.25.1
python-dotenv==0.17.0